添加了获取成员信息和批量获取成员信息的初始化版本

+++ app/common/type.go
@@ -95,8 +95,8 @@ type GetMemberResponse struct {
 // 批量获取成员信息
 
 type GetMemberListRequest struct {
-	Offset int
-	Limit  int
+	Offset int64
+	Limit  int64
 }
 
 type GetMemberListResponse struct {
